requirements/local.txt
humanfriendly==10.0
Pillow==8.3.2
pyinstrument==4.0.3
